refactor file structure for rule editor
